module5/bruin/zoomcamp/pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
requests==2.31.0
python-dateutil==2.9.0.post0
pyarrow==16.0.0